refactor: move App.vue to App/ folder
